9773 Update tests to include location authorization
